# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Taxi App - Display users' location"
  },
  "servers": [
    {
      "url": "https://secret-mesa-70576.herokuapp.com/"
    }
  ],
  "paths": {
...
    "/signup": {
      "post": {
        "summary": "Add a user",
        "tags": ["users"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description":
              "The username and the password of the user. Usernames must be unique and cannot be blank",
            "schema": {
              "example": {
                "username": "user01",
                "passsword": "password01"
              },
              "type": "JSON"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Null response"
          },
          "500": {
            "description":
              "'User validation failed: username: should be unique' or 'User validation failed: username: cannot be blank'"
          }
        }
      }
    },
...
    "/{username}/logout": {
      "put": {
        "summary":
          "Change the availability of the user to unavailable",
        "tags": ["users"],
        "parameters": [
          {
            "name": "{username}/logout",
            "in": "path",
            "required": true,
            "description": "Change the availability of the user to unavailable",
            "schema": {
              "example": {
                "username": "newUser01",
                "passsword": "new_password01"
              },
              "type": "path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Null response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}